requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.3.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.5.4
mysqlclient==2.2.0
packaging==25.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
